+++ SocialNetwork/SocialNetwork.Dal/Models/AttachmentPost.cs
@@ -4,8 +4,8 @@ namespace SocialNetwork.Dal.Models
 {
     public class AttachmentPost
     {
-        public BigInteger PostId { get; set; }
-        public BigInteger AttachmentId { get; set; }
+        public int PostId { get; set; }
+        public int AttachmentId { get; set; }
 
         public virtual Attachment Attachment { get; set; }
         public virtual Post Post { get; set; }
