+++ SocialNetwork/SocialNetwork.Dal/Models/Post.cs
@@ -13,7 +13,7 @@ namespace SocialNetwork.Dal.Models
             AttachmentPost = new HashSet<AttachmentPost>();
         }
 
-        public BigInteger Id { get; set; }
+        public int Id { get; set; }
         public string Text { get; set; }
         public Guid UserId { get; set; }
         public bool IsArchived { get; set; }
