+++ SocialNetwork/SocialNetwork.Dal/ViewModels/In/ReactionPostViewModel.cs
@@ -7,7 +7,7 @@ namespace SocialNetwork.Dal.ViewModels.In
     public class ReactionPostViewModel
     {
         [Required]
-        public BigInteger PostId { get; set; }
+        public int PostId { get; set; }
         public Guid ReactionId { get; set; }
     }
 }
